refactor: shift deleteForm to UpdateFormService
