+++ src/public/services/__tests__/UpdateFormService.test.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { ObjectId } from 'bson'
+import { StatusCodes } from 'http-status-codes'
 import MockAxios from 'jest-mock-axios'
 
 import { FormUpdateParams } from 'src/app/modules/form/admin-form/admin-form.types'
@@ -13,6 +14,7 @@ import {
 import * as SubmissionUtil from '../../utils/submission'
 import {
   ADMIN_FORM_ENDPOINT,
+  deleteForm,
   submitEmailModeFormPreview,
   submitStorageModeFormPreview,
   transferOwner,
@@ -157,6 +159,42 @@ describe('UpdateFormService', () => {
         MOCK_CONTENT,
         // Should default to stringified null
         { params: { captchaResponse: 'null' } },
+      )
+    })
+  })
+
+  describe('deleteForm', () => {
+    it('should successfully call delete endpoint', async () => {
+      // Arrange
+      const MOCK_FORM_ID = new ObjectId().toHexString()
+
+      // Act
+      const actualPromise = deleteForm(MOCK_FORM_ID)
+      MockAxios.mockResponse({
+        status: StatusCodes.OK,
+        data: { message: 'Form has been archived' },
+      })
+      await actualPromise
+
+      // Assert
+      expect(MockAxios.delete).toHaveBeenCalledWith(
+        `${ADMIN_FORM_ENDPOINT}/${MOCK_FORM_ID}`,
+      )
+    })
+
+    it('should reject with error message if DELETE request fails', async () => {
+      // Arrange
+      const expected = new Error('error')
+      const MOCK_FORM_ID = new ObjectId().toHexString()
+
+      // Act
+      const actualPromise = deleteForm(MOCK_FORM_ID)
+      MockAxios.mockError(expected)
+
+      await expect(actualPromise).rejects.toEqual(expected)
+      // Assert
+      expect(MockAxios.delete).toHaveBeenCalledWith(
+        `${ADMIN_FORM_ENDPOINT}/${MOCK_FORM_ID}`,
       )
     })
   })
